Move runtime tests under `java-tests`
