--- java-runtime/src/test/resources/desc.ref
@@ -1,1 +1,0 @@
-known_types_test.desc
