+++ java-tests/runtime/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,36 +24,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validate
+import io.spine.internal.dependency.Spine
 
-import io.kotest.matchers.types.shouldBeSameInstanceAs
-import io.spine.test.validate.Meal
-import io.spine.test.validate.Meat
-import io.spine.test.validate.meal
-import org.junit.jupiter.api.Disabled
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-
-@DisplayName("Validation extensions for `Message` should")
-@Disabled("Until Validation migrates to new ProtoData")
-internal class MessageExtensionsSpec {
-
-    @Nested
-    inner class `Check if the message is valid` {
-
-        @Test
-        fun `returning 'this' if so`() {
-            val meal = meal { meat = Meat.getDefaultInstance() }
-            meal.checkValid() shouldBeSameInstanceAs meal
-        }
-
-        @Test
-        fun `throwing 'ValidationException' if not`() {
-            assertThrows<ValidationException> {
-                Meal.getDefaultInstance().checkValid()
-            }
-        }
-    }
+dependencies {
+    testImplementation(Spine.Logging.lib)
+    testImplementation(Spine.testlib)
 }
