+++ java-tests/runtime/src/test/java/io/spine/validate/FieldAwareMessageTest.java
@@ -24,19 +24,36 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-buildscript {
-    forceCodegenPlugins()
-}
+package io.spine.validate;
 
-/*
- * Disable the generation of rejections, since we don't want
- * other plugins (potentially, still not using the latest ProtoData API)
- * to interfere with the tests of Validation ProtoData plugin.
- */
-modelCompiler {
-    java {
-        codegen {
-            rejections().enabled.set(false)
-        }
+import com.google.common.truth.Truth;
+import io.spine.test.validate.AggregateState;
+import io.spine.validate.given.FieldAwareMessageTestEnv;
+import io.spine.validate.given.FieldAwareMessageTestEnv.BrokenFieldAware;
+import io.spine.validate.given.FieldAwareMessageTestEnv.FieldAwareMsg;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static com.google.common.truth.Truth.assertThat;
+import static io.spine.testing.Assertions.assertIllegalArgument;
+import static io.spine.validate.given.FieldAwareMessageTestEnv.msg;
+
+@DisplayName("`FieldAwareMessage` should")
+class FieldAwareMessageTest {
+
+    @Test
+    @DisplayName("read values when `readValues` is properly implemented")
+    void readValues() {
+        var msg = FieldAwareMessageTestEnv.msg();
+        var state = new FieldAwareMsg(msg);
+        Truth.assertThat(state.checkFieldsReachable()).isTrue();
+    }
+
+    @Test
+    @DisplayName("fail to read values when `readValues` has implementation issues")
+    void failToReadValues() {
+        var msg = FieldAwareMessageTestEnv.msg();
+        var state = new BrokenFieldAware(msg);
+        assertIllegalArgument(state::checkFieldsReachable);
     }
 }
