+++ java-tests/runtime/src/test/java/io/spine/validate/IfMissingErrorMsgSpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -43,7 +43,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
 @DisplayName(VALIDATION_SHOULD + "use `if_missing` field option")
-@Disabled("Until Validation migrates to new ProtoData")
 internal class IfMissingErrorMsgSpec : ValidationOfConstraintTest() {
 
     @Nested inner class
