+++ java-tests/runtime/src/test/java/io/spine/validate/OneofSpec.kt
