+++ java-tests/runtime/src/test/java/io/spine/validate/ValidateUntilitySpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -97,7 +97,7 @@ internal class ValidateUntilitySpec : UtilityClassTest<Validate>(Validate::class
                 .setId("JC 424")
                 .setBirthplace("Edinburgh")
                 .build()
-            checkViolated(oldValue, newValue, Companion.BIRTHPLACE)
+            checkViolated(oldValue, newValue, BIRTHPLACE)
         }
 
         @Test
