+++ java-tests/runtime/src/test/java/io/spine/validate/given/FieldAwareMessageTestEnv.java
