+++ java-tests/runtime/src/test/java/io/spine/validate/given/package-info.java
