+++ java-tests/runtime/src/test/java/io/spine/validate/option/RequiredSpec.kt
