+++ java-tests/runtime/src/test/kotlin/io/spine/validate/AnyValidationSpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -40,7 +40,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
 @DisplayName(VALIDATION_SHOULD + "when validating `google.protobuf.Any`")
-@Disabled("Until Validation migrates to new ProtoData")
 internal class AnyValidationSpec : ValidationOfConstraintTest() {
 
     /**
