+++ java-tests/runtime/src/test/kotlin/io/spine/validate/EnclosedMessageValidationSpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,7 +46,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 
 @DisplayName(VALIDATION_SHOULD + "validate enclosed messages and")
-@Disabled("Until Validation migrates to new ProtoData")
 internal class EnclosedMessageValidationSpec : ValidationOfConstraintTest() {
 
     @Test
