+++ java-tests/runtime/src/test/kotlin/io/spine/validate/ValidationOfConstraintTest.kt
