+++ java-tests/runtime/src/test/kotlin/io/spine/validate/option/GoesSpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,14 +35,12 @@ import io.spine.test.validate.paymentId
 import io.spine.test.validate.withFieldNotFound
 import io.spine.validate.ValidationOfConstraintTest
 import io.spine.validate.ValidationOfConstraintTest.Companion.VALIDATION_SHOULD
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
 @DisplayName(VALIDATION_SHOULD + "analyze `(goes)` option and find out that ")
-@Disabled("Until Validation migrates to new ProtoData")
 internal class GoesSpec : ValidationOfConstraintTest() {
 
     @Nested
