+++ java-tests/runtime/src/test/proto/spine/test/validation/alt_fields.proto
