+++ java-tests/runtime/src/test/proto/spine/test/validation/any_fields.proto
