+++ java-tests/runtime/src/test/proto/spine/test/validation/commands.proto
