+++ java-tests/runtime/src/test/proto/spine/test/validation/distinct_option_test.proto
