+++ java-tests/runtime/src/test/proto/spine/test/validation/field_value_test.proto
