+++ java-tests/runtime/src/test/proto/spine/test/validation/goes_option_test.proto
