+++ java-tests/runtime/src/test/proto/spine/test/validation/if_missing_test.proto
