+++ java-tests/runtime/src/test/proto/spine/test/validation/is_required_test.proto
