+++ java-tests/runtime/src/test/proto/spine/test/validation/messages.proto
