+++ java-tests/runtime/src/test/proto/spine/test/validation/min_max_test.proto
