+++ java-tests/runtime/src/test/proto/spine/test/validation/oneof.proto
