+++ java-tests/runtime/src/test/proto/spine/test/validation/pattern_option_test.proto
