+++ java-tests/runtime/src/test/proto/spine/test/validation/patterns.proto
