+++ java-tests/runtime/src/test/proto/spine/test/validation/range_option_test.proto
