+++ java-tests/runtime/src/test/proto/spine/test/validation/required_field_test.proto
