+++ java-tests/runtime/src/test/proto/spine/test/validation/required_test.proto
