+++ java-tests/runtime/src/test/proto/spine/test/validation/set_once_test.proto
