+++ java-tests/runtime/src/test/proto/spine/test/validation/validate_test.proto
