+++ java-tests/runtime/src/test/proto/spine/test/validation/validation_constraints_test.proto
